# gen/openapi.deref.yaml
openapi: 3.0.3
info:
  title: Financial-grade API Connector
  description: >-
    Use this companion service to register with and call OpenID providers.


    Optionally, it can be used in conjunction with the [Platern
    Directory](https://platern.com) in [Open
    Banking](https://opendata.substack.com/p/an-overview-of-open-banking)
    applications.


    This service uses a forked, open-source [Node OpenID
    Client](https://github.com/dextersjab/node-openid-client).


    Similarly, it uses the MIT licence, allowing the most open usage possible,
    whether for:

    - making modifications

    - commercial purposes

    - private usage

    - distributing
        
    Learn more:

    - [Platern](https://platern.com)

    - [GitHub](https://github.com/platern/fapi-connector)


    Feel free to raise issues
    [here](https://github.com/platern/fapi-connector/issues).
  termsOfService: 'https://platern.com/terms'
  contact:
    email: hello@platern.com
  license:
    name: MIT
    url: 'https://github.com/platern/openid-client/LICENCE'
  version: '0.1'
paths:
...
  /authorization:
    get:
      tags:
        - Authorization
      description: |-
        Generate an authorization URL for the data owner to grant access.

        The query parameters you need to send to this endpoint depend on
        how you're using *fapi-connector*.

        If you're using Platern Web, use:
          - `providerID`

        If you're connecting without Platern Web (e.g. to Open Banking), use:
          - `oauth2Grant`
          - `oauth2GrantRequest`
      summary: Generate an authorization URL for the data owner to grant access.
      parameters:
        - in: query
          description: Use the `clientID` registered with your API provider.
          name: client
          required: true
          schema:
            type: string
          examples:
            '200':
              value: XXX
        - in: query
          description: 'If you''re using Platern Web, pass the `provider`.'
          name: provider
          schema:
            type: string
          examples:
            '200':
              value: XXX
        - in: query
          description: |-
            Currently supports the following Open Banking UK specifications:
              - `obieAis`
              - `obieCof`
              - `obieDcr`
              - `obieEvent`
              - `obiePis`
              - `obieProduct`
          name: specification
          required: true
          schema:
            type: string
          examples:
            '200':
              value: obieAis
        - in: query
          description: |-
            #### OAuth2: `state`

            RECOMMENDED.  An opaque value used by the client to maintain
                     state between the request and callback.  The authorization
                     server includes this value when redirecting the user-agent back
                     to the client.  The parameter SHOULD be used for preventing
                     cross-site request forgery as described in [RFC 6749 Section 10.12](https://www.rfc-editor.org/rfc/rfc6749#section-10.12).
          name: oauth2State
          required: true
          schema:
            type: string
          examples:
            '200':
              value: cd444cb3-68b8-42b4-a9dd-aa3c75ccade2
        - in: query
          name: openIDNonce
          required: true
          schema:
            type: string
          examples:
            '200':
              value: 96d3dd10-2dc7-4b93-835d-89ddd8665b06
        - in: query
          description: >-
            Required for Open Banking UK. Usually has a path ending with
            `/account-access-consents`.


            Also see [FAPI Grant Management extension to
            OAuth2](https://openid.net/specs/fapi-grant-management-01.html).
          name: oauth2GrantUrl
          schema:
            type: string
          examples:
            '200':
              summary: 200 Ozone Sandbox
              value: >-
                https://ob19-rs1.o3bank.co.uk:4501/open-banking/v3.1/aisp/account-access-consents
        - in: query
          description: >-
            Base64-encoded grant request payload.

            Required for Open Banking UK. Usually has a path ending with
            `/account-access-consents`.


            Also see [FAPI Grant Management extension to
            OAuth2](https://openid.net/specs/fapi-grant-management-01.html).
          name: oauth2GrantRequest
          schema:
            type: string
          examples:
            '200':
              value: >-
                eyJEYXRhIjp7IlBlcm1pc3Npb25zIjpbIlJlYWRBY2NvdW50c0RldGFpbCIsIlJlYWRCYWxhbmNlcyIsIlJlYWRCZW5lZmljaWFyaWVzRGV0YWlsIiwiUmVhZERpcmVjdERlYml0cyIsIlJlYWRQcm9kdWN0cyIsIlJlYWRTdGFuZGluZ09yZGVyc0RldGFpbCIsIlJlYWRUcmFuc2FjdGlvbnNDcmVkaXRzIiwiUmVhZFRyYW5zYWN0aW9uc0RlYml0cyIsIlJlYWRUcmFuc2FjdGlvbnNEZXRhaWwiLCJSZWFkT2ZmZXJzIiwiUmVhZFBBTiIsIlJlYWRQYXJ0eSIsIlJlYWRQYXJ0eVBTVSIsIlJlYWRTY2hlZHVsZWRQYXltZW50c0RldGFpbCIsIlJlYWRTdGF0ZW1lbnRzRGV0YWlsIl0sIkV4cGlyYXRpb25EYXRlVGltZSI6IjIwMjMtMTItMDFUMDA6MDA6MDArMDA6MDAiLCJUcmFuc2FjdGlvbkZyb21EYXRlVGltZSI6IjIwMTYtMTAtMTVUMDA6MDA6MDArMDA6MDAiLCJUcmFuc2FjdGlvblRvRGF0ZVRpbWUiOiIyMDIyLTEyLTE1VDAwOjAwOjAwKzAwOjAwIn0sIlJpc2siOnt9fQ==
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: &ref_9
                type: object
                properties:
                  url:
                    type: string
                required:
                  - url
              examples:
                '200':
                  value:
                    url: >-
                      https://auth.abcbank.com/authorize?client_id=CLIENT_ID&scope=openid&response_type=code&redirect_uri=https%3A%2F%2Fexample.com%2Foauth%2Fcallback%2F&request=eyJhbGciOiJQUzI1NiIsInR5cCI6Im9hdXRoLWF1dGh6LXJlcStqd3QifQ.eyJzY29wZSI6ImFjY291bnQiLCJzdGF0ZSI6IjgxMDVlYWFkLTY4MzktNGZkMC05MWZhLTAzMThlZTYyZGE5YiIsIm5vbmNlIjoiOTAxMjlkYzAtOWNiNy00ZmMzLTk2ZmQtMzQ4ZjNiMWIxMzgzIiwibWF4X2FnZSI6ODY0MDAsInJlZGlyZWN0X3VyaSI6Imh0dHBzOi8vZXhhbXBsZS5jb20vb2F1dGgvY2FsbGJhY2svIiwiY2xhaW1zIjp7InVzZXJpbmZvIjp7Im9wZW5iYW5raW5nX2ludGVudF9pZCI6eyJ2YWx1ZSI6IkNVU1RPTUVSX0FDQ09VTlRfQ09OU0VOVCIsImVzc2VudGlhbCI6dHJ1ZX19LCJpZF90b2tlbiI6eyJvcGVuYmFua2luZ19pbnRlbnRfaWQiOnsidmFsdWUiOiJDVVNUT01FUl9BQ0NPVU5UX0NPTlNFTlQiLCJlc3NlbnRpYWwiOnRydWV9LCJhY3IiOnsiZXNzZW50aWFsIjp0cnVlLCJ2YWx1ZXMiOlsidXJuOm9wZW5iYW5raW5nOnBzZDI6c2NhIiwidXJuOm9wZW5iYW5raW5nOnBzZDI6Y2EiXX19fSwiaXNzIjoiQ0xJRU5UX0lEIiwiY2xpZW50X2lkIjoiQ0xJRU5UX0lEIiwianRpIjoiNW9HQzd1UUZsRndDQWEwZFFBbHg1a0xwbEx2RjJ6aGlMbjMxQVVhVHFYYyIsImlhdCI6MTY2NjE2MzAzNSwiZXhwIjoxNjY2MTYzMzM1LCJuYmYiOjE2NjYxNjMwMzV9.RYHl8EbDPddwOnA_dNRvjprd5eVn2v5E4Db7q5JHhaULYGXHhShsDfBGbJrKieZgAm3IHk6borr-Pl5P2qK8gBsuPLAGYONT2ZABD5ktAEksprWBmt0SMDBB8URWnabuBgYYmErLk2k8_QpVhVtDEwl_q-8CQfH1rn8-k7M3xPF7egZY7nLd7sIOw8-EOEWglUzO_irrdRplZIyFMA91pvDBQuPSjMvFTTfOcXXOkWKngQHdVPdw3JJuHgF3kmMNA0b0sFlPvYFobnCPEDDEc2MMqJ9RmdAZ0ENiFopLs_3V8_pWXo4c3OcmlD36-qxEgkLhKSCpBfo464rb3MaYXw
  /token:
    description: >-
      Proxy service to perform an OAuth2 token request with an API provider for
      an access token.
    post:
      tags:
        - Authorization
      description: >-
        Make an OAuth2 [token
        request](https://www.rfc-editor.org/rfc/rfc6749#section-4.1.3) to get
        access credentials.
      summary: Access token request.
      parameters:
        - in: query
          description: Use the `clientID` registered with your API provider.
          name: client
          schema:
            type: string
          examples:
            '200':
              value: XXX
      requestBody:
        content:
          application/json:
            schema: &ref_10
              type: object
              properties:
                oauth2State:
                  description: OAuth2 state
                  type: string
                oauth2Code:
                  description: OAuth2 code
                  type: string
                openIDNonce:
                  description: OpenID nonce
                  type: string
            examples:
              '200':
                summary: 200 (w/ Platern Web)
                value:
                  oauth2State: cd444cb3-68b8-42b4-a9dd-aa3c75ccade2
                  oauth2Code: 295f8dfa-11f1-4424-9c8a-c498e0e6b8cd
                  openIDNonce: 96d3dd10-2dc7-4b93-835d-89ddd8665b06
              x-200.WithOIDD:
                summary: 200 (w/ OpenID Discovery URL)
                value:
                  oauth2State: cd444cb3-68b8-42b4-a9dd-aa3c75ccade2
                  oauth2Code: 295f8dfa-11f1-4424-9c8a-c498e0e6b8cd
                  openIDNonce: 96d3dd10-2dc7-4b93-835d-89ddd8665b06
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: &ref_11
                type: object
                properties:
                  oauth2TokenSet: &ref_12
                    title: OAuth2 Token Response
                    type: object
                    properties:
                      access_token:
                        description: The access token issued by the authorization server.
                        type: string
                      expires_in:
                        description: >-
                          RECOMMENDED.  The lifetime in seconds of the access
                          token.  For

                          example, the value "3600" denotes that the access
                          token will

                          expire in one hour from the time the response was
                          generated.

                          If omitted, the authorization server SHOULD provide
                          the

                          expiration time via other means or document the
                          default value.
                        type: integer
                      refresh_token:
                        description: >-
                          Can be used to obtain new access tokens using the same

                          authorization grant as described in Section 6 of RFC
                          6749.
                        type: string
                      scope:
                        description: >-
                          The authorization server uses the "scope" response
                          parameter to

                          inform the client of the scope of the access token
                          issued.


                          The value of the scope parameter is expressed as a
                          list of space-

                          delimited, case-sensitive strings.  The strings are
                          defined by the

                          authorization server.  If the value contains multiple
                          space-delimited

                          strings, their order does not matter, and each string
                          adds an

                          additional access range to the requested scope.
                        type: string
                      token_type:
                        description: |-
                          The type of the token issued as described in
                          Section 7.1.  Value is case insensitive.
                        type: string
                      state:
                        description: >-
                          REQUIRED if the "state" parameter was present in the
                          client
                                   authorization request.  The exact value received from the
                                   client.
                        type: string
                    required:
                      - access_token
                      - expires_in
                      - token_type
                required:
                  - oauth2AccessToken
              examples:
                '200':
                  value:
                    tokenSet: null
...